requirements.txt
speedtest-cli==2.1.3
requests==2.31.0
python-dotenv==0.19.2
cryptography==39.0.1